GUACAMOLE-195: Rename extension to guacamole-auth-header to more accurately reflect nature of the authentication.
