+++ extensions/guacamole-auth-header/.gitignore
