+++ extensions/guacamole-auth-header/src/licenses/DISCLAIMER
