+++ extensions/guacamole-auth-header/src/licenses/LICENSE
