+++ extensions/guacamole-auth-header/src/licenses/NOTICE
