+++ extensions/guacamole-auth-header/src/licenses/bundled/README
