+++ extensions/guacamole-auth-header/src/licenses/bundled/aopalliance-1.0/LICENSE
