+++ extensions/guacamole-auth-header/src/licenses/bundled/guice-3.0/COPYING
