+++ extensions/guacamole-auth-header/src/main/assembly/dist.xml
