+++ extensions/guacamole-auth-header/src/main/resources/guac-manifest.json
@@ -3,10 +3,10 @@
     "guacamoleVersion" : "0.9.11-incubating",
 
     "name"      : "HTTP Header Authentication Extension",
-    "namespace" : "guac-http",
+    "namespace" : "guac-header",
 
     "authProviders" : [
-        "org.apache.guacamole.auth.http.HTTPAuthenticationProvider"
+        "org.apache.guacamole.auth.header.HTTPHeaderAuthenticationProvider"
     ]
 
 }
