+++ pom.xml
@@ -50,7 +50,7 @@
 
         <!-- Authentication extensions -->
         <module>extensions/guacamole-auth-duo</module>
-        <module>extensions/guacamole-auth-http</module>
+        <module>extensions/guacamole-auth-header</module>
         <module>extensions/guacamole-auth-jdbc</module>
         <module>extensions/guacamole-auth-ldap</module>
         <module>extensions/guacamole-auth-noauth</module>
